# wordpress/themes/cdcf-headless/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CDCF CMS API",
    "description": "REST API for the Catholic Digital Commons Foundation headless CMS. Includes WordPress custom endpoints (cdcf/v1) for content management and translation, plus Next.js API routes for preview, cache revalidation, and sitemaps.",
    "version": "1.0.0",
    "contact": {
      "name": "Catholic Digital Commons Foundation",
      "url": "https://catholicdigitalcommons.org"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/wp-json",
      "description": "Local WordPress (Docker)"
    },
    {
      "url": "https://cms.catholicdigitalcommons.org/wp-json",
      "description": "Production WordPress"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Local Next.js"
    },
    {
      "url": "https://staging.catholicdigitalcommons.org/api",
      "description": "Production Next.js"
    }
  ],
...
  "paths": {
...
    "/refer-local-group": {
      "post": {
        "tags": ["Referral"],
        "summary": "Proxy local group referral to WordPress (Next.js)",
        "description": "Next.js API route that validates a honeypot field, applies in-memory rate limiting (3/hour/IP), and proxies the submission to the WordPress `cdcf/v1/refer-local-group` endpoint.",
        "operationId": "referLocalGroupProxy",
        "security": [],
        "servers": [
          { "url": "http://localhost:3000/api", "description": "Local Next.js" },
          { "url": "https://staging.catholicdigitalcommons.org/api", "description": "Production Next.js" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["group_name", "description", "url", "submitter_name", "submitter_email"],
                "properties": {
                  "group_name": { "type": "string" },
                  "description": { "type": "string" },
                  "url": { "type": "string", "format": "uri" },
                  "location": { "type": "string" },
                  "submitter_name": { "type": "string" },
                  "submitter_email": { "type": "string", "format": "email" },
                  "website": {
                    "type": "string",
                    "description": "Honeypot field — if filled, request silently succeeds without creating a post"
                  },
                  "elapsed_ms": {
                    "type": "number",
                    "description": "Milliseconds since the modal was opened. Submissions under 3000ms are silently discarded (bot detection)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Referral submitted (or honeypot triggered)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "post_id": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields or invalid body"
          },
          "429": {
            "description": "Rate limited"
          },
          "502": {
            "description": "Could not reach WordPress backend"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "WordPress Application Password. Username is the WordPress username; password is the application password (spaces are normal)."
      },
      "previewSecret": {
        "type": "apiKey",
        "in": "query",
        "name": "secret",
        "description": "Shared secret matching the WP_PREVIEW_SECRET environment variable."
      }
    },
...
  }
}